[Add]マルチスレッド処理を有効化など

+++ IGS2ONS_Noesis-loveesm.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python3
 from PIL import Image, ImageEnhance
+import concurrent.futures
 import subprocess
 import shutil
 import glob
@@ -7,24 +8,37 @@ import sys
 import os
 import re
 
-same_hierarchy = (os.path.dirname(sys.argv[0]))#同一階層のパスを変数へ代入
-EXE_IGS = os.path.join(same_hierarchy,'igscriptD.exe')
-DEFAULT_TXT = os.path.join(same_hierarchy,'default.txt')
 
-#same_hierarchy = os.path.join(same_hierarchy,'Gungnir_loveesm_EXT')#debug
-DIR_SCR = os.path.join(same_hierarchy,'script')
-DIR_SCR_DEC = os.path.join(same_hierarchy,'script_dec')
+def img_resize_main(p, DIR_IMG):
+	p_name = os.path.basename(p)
+	p_r = os.path.join(DIR_IMG, p_name)
 
-DIR_BG = os.path.join(same_hierarchy,'bgimage')
-DIR_FG = os.path.join(same_hierarchy,'fgimage')
-DIR_SYS = os.path.join(same_hierarchy,'system')
-DIR_IMG = os.path.join(same_hierarchy,'image')
+	im = Image.open(p)
+	w = int(im.width * 600 / 768)
+	h = int(im.height * 600 / 768)
 
-jmp_cnt = 0
-jmp_list = []
+	im_r = im.resize((w, h), Image.Resampling.LANCZOS)
+
+	#画像個別処理
+	if str(p_name).lower() == 'window_base.png':
+		im_r.putalpha(192)
+
+	elif str(p_name).lower() == 'load_base.png':
+		enhancer = ImageEnhance.Brightness(im_r)
+		im_r = enhancer.enhance(0.4)
+
+	elif str(p_name).lower() == 'config_base.png':
+		enhancer = ImageEnhance.Brightness(im_r)
+		im_r = enhancer.enhance(0.2)
+
+	elif str(p_name).lower() == 'view_cg.png':
+		enhancer = ImageEnhance.Brightness(im_r)
+		im_r = enhancer.enhance(0.2)
+
+	im_r.save(p_r)
 
 
-def img_resize():
+def img_resize(DIR_BG, DIR_FG, DIR_SYS, DIR_IMG):
 	l = []
 	l += glob.glob(os.path.join(DIR_BG, '*.png'))
 	l += glob.glob(os.path.join(DIR_FG, '*.png'))
@@ -32,36 +46,27 @@ def img_resize():
 
 	os.makedirs(DIR_IMG, exist_ok=True)
 
-	for p in l:
-		p_name = os.path.basename(p)
-		p_r = os.path.join(DIR_IMG, p_name)
+	#本処理 - マルチスレッドで高速化
+	with concurrent.futures.ThreadPoolExecutor() as executor:
+		futures = []
+		for p in l: futures.append(executor.submit(img_resize_main, p, DIR_IMG))
+		concurrent.futures.as_completed(futures)
 
-		im = Image.open(p)
-		w = int(im.width * 600 / 768)
-		h = int(im.height * 600 / 768)
-
-		im_r = im.resize((w, h), Image.Resampling.LANCZOS)
-
-		#画像個別処理
-		if str(p_name).lower() == 'window_base.png':
-			im_r.putalpha(192)
-
-		elif str(p_name).lower() == 'load_base.png':
-			enhancer = ImageEnhance.Brightness(im_r)
-			im_r = enhancer.enhance(0.4)
-
-		elif str(p_name).lower() == 'config_base.png':
-			enhancer = ImageEnhance.Brightness(im_r)
-			im_r = enhancer.enhance(0.2)
-
-		elif str(p_name).lower() == 'view_cg.png':
-			enhancer = ImageEnhance.Brightness(im_r)
-			im_r = enhancer.enhance(0.2)
-
-		im_r.save(p_r)
+	#クレジット結合
+	im1 = Image.open(os.path.join(DIR_IMG, 'credit_01.png'))
+	im2 = Image.open(os.path.join(DIR_IMG, 'credit_02.png'))
+	im3 = Image.open(os.path.join(DIR_IMG, 'credit_03.png'))
+	imnew = Image.new('RGBA', (im1.width, im1.height+im2.height+im3.height))
+	imnew.paste(im1, (0, 0))
+	imnew.paste(im2, (0, im1.height))
+	imnew.paste(im3, (0, im1.height+im2.height))
+	imnew.save(os.path.join(DIR_IMG, 'credit_join.png'))
 
 
-def text_dec():
+
+
+
+def text_dec(DIR_SCR_DEC, DIR_SCR, EXE_IGS):
 	os.makedirs(DIR_SCR_DEC, exist_ok=True)
 
 	for p in glob.glob(os.path.join(DIR_SCR, '*.s')):
@@ -72,8 +77,10 @@ def text_dec():
 		shutil.move(p, DIR_SCR_DEC)
 
 
-def text_cnv():
-	global jmp_cnt
+def text_cnv(DEFAULT_TXT, DIR_SCR_DEC, ZERO_TXT):
+	jmp_cnt = 0
+	jmp_list = []
+
 	with open(DEFAULT_TXT) as f:
 		txt = f.read()
 
@@ -202,7 +209,7 @@ def text_cnv():
 	txt = txt.replace(r';JMP_CNT12_', r'select "'+jmp_list[11]+r'",*S6A,"'+jmp_list[12]+r'",*S6B'+'\n*S6B')
 	txt = txt.replace(r'goto *SCR_A0420', 'goto *SCR_A0420\n*S6A')
 	txt = txt.replace(r';SCR_A0660_END', 'bgmstop:dwavestop -1:csp -1:reset')
-	txt = txt.replace(r';SCR_A0440Z_END', 'bgmstop:dwavestop -1:skipoff:mov %200,1:bgmonce "bgm\\Led_a.ogg":bg "image\\ed1_ayaka.png",10:click:bgmstop:dwavestop -1:csp -1:reset')#ED1
+	txt = txt.replace(r';SCR_A0440Z_END', 'bgmstop:dwavestop -1:skipoff:mov %200,1:bgmonce "bgm\\Led_a.ogg":bg "image\\ed1_ayaka.png",10:mov %150,93066:gosub *staffroll:click:bgmstop:dwavestop -1:csp -1:reset')#ED1
 	#選択肢7
 	txt = txt.replace(r';JMP_CNT14_', r'select "'+jmp_list[13]+r'",*S7A,"'+jmp_list[14]+r'",*S7B'+'\n*S7A')
 	txt = txt.replace(r'goto *SCR_B0540A', 'goto *SCR_B0540A\n*S7B')
@@ -211,7 +218,7 @@ def text_cnv():
 	txt = txt.replace(r';JMP_CNT16_', r'select "'+jmp_list[15]+r'",*S8A,"'+jmp_list[16]+r'",*S8B'+'\n*S8B')
 	txt = txt.replace(r'goto *SCR_B0630', 'goto *SCR_B0630\n*S8A')
 	txt = txt.replace(r';SCR_b0680z_END', 'bgmstop:dwavestop -1:csp -1:reset')
-	txt = txt.replace(r';SCR_B0640Z_END', 'bgmstop:dwavestop -1:skipoff:mov %201,1:bgmonce "bgm\\Led_a.ogg":bg "image\\ed2_ayaka.png",10:click:bgmstop:dwavestop -1:csp -1:reset')#ED2
+	txt = txt.replace(r';SCR_B0640Z_END', 'bgmstop:dwavestop -1:skipoff:mov %201,1:bgmonce "bgm\\Led_a.ogg":bg "image\\ed2_ayaka.png",10:mov %150,93066:gosub *staffroll:click:bgmstop:dwavestop -1:csp -1:reset')#ED2
 	#選択肢9
 	txt = txt.replace(r';JMP_CNT18_', r'select "'+jmp_list[17]+r'",*S9A,"'+jmp_list[18]+r'",*S9B'+'\n*S9A\nmov %101,1')
 	txt = txt.replace('の準備を始めた。\\', 'の準備を始めた。\\\ntati_reset')
@@ -219,12 +226,12 @@ def text_cnv():
 	txt = txt.replace('着てもらう。\\', '着てもらう。\\\nif %101==2 goto *S9SKIP')
 	txt = txt.replace('かは顔を見合わせ、それからにっこりと笑って言った。\\', 'かは顔を見合わせ、それからにっこりと笑って言った。\\\ngoto *S9END\n*S9SKIP')
 	txt = txt.replace('華は顔を見合わせ、それからにっこりと笑って言った。\\', '華は顔を見合わせ、それからにっこりと笑って言った。\\\n*S9END')
-	txt = txt.replace(';SCR_C0700Z_END', 'tati_reset:csp -1:bgmstop:dwavestop -1:skipoff:mov %202,1:bgmonce "bgm\\Led_s.ogg":bg "image\\ed3_harem.png",10:click:bgmstop:dwavestop -1:csp -1:reset')#ED3
+	txt = txt.replace(';SCR_C0700Z_END', 'tati_reset:csp -1:bgmstop:dwavestop -1:skipoff:mov %202,1:bgmonce "bgm\\Led_s.ogg":bg "image\\ed3_harem.png",10:mov %150,126007:gosub *staffroll:click:bgmstop:dwavestop -1:csp -1:reset')#ED3
 
-	open(os.path.join(same_hierarchy,'0.txt'), 'w', errors='ignore').write(txt)
+	open(ZERO_TXT, 'w', errors='ignore').write(txt)
 
 
-def file_check():
+def file_check(EXE_IGS, DEFAULT_TXT, DIR_SCR, DIR_BG, DIR_FG, DIR_SYS):
 	c = True
 	for p in [EXE_IGS, DEFAULT_TXT, DIR_SCR, DIR_BG, DIR_FG, DIR_SYS]:
 		if not os.path.exists(p):
@@ -234,17 +241,39 @@ def file_check():
 	return c
 
 
-def junk_del():
+def junk_del(DIR_BG, DIR_FG, DIR_SYS, DIR_SCR, DIR_SCR_DEC, DEFAULT_TXT, debug):
 	shutil.rmtree(DIR_BG)
 	shutil.rmtree(DIR_FG)
 	shutil.rmtree(DIR_SYS)
 	shutil.rmtree(DIR_SCR)
 	shutil.rmtree(DIR_SCR_DEC)
-	os.remove(DEFAULT_TXT)
+	if not debug: os.remove(DEFAULT_TXT)
 
 
-if file_check():
-	img_resize()
-	text_dec()
-	text_cnv()
-	junk_del()
+def main():
+
+	#デバッグ
+	debug = 0
+
+	same_hierarchy = (os.path.dirname(sys.argv[0]))#同一階層のパスを変数へ代入
+	EXE_IGS = os.path.join(same_hierarchy,'igscriptD.exe')
+	DEFAULT_TXT = os.path.join(same_hierarchy,'default.txt')
+
+	if debug: same_hierarchy = os.path.join(same_hierarchy,'Gungnir_loveesm_EXT')#debug
+	ZERO_TXT = os.path.join(same_hierarchy,'0.txt')
+	DIR_SCR = os.path.join(same_hierarchy,'script')
+	DIR_SCR_DEC = os.path.join(same_hierarchy,'script_dec')
+
+	DIR_BG = os.path.join(same_hierarchy,'bgimage')
+	DIR_FG = os.path.join(same_hierarchy,'fgimage')
+	DIR_SYS = os.path.join(same_hierarchy,'system')
+	DIR_IMG = os.path.join(same_hierarchy,'image')
+
+	if file_check(EXE_IGS, DEFAULT_TXT, DIR_SCR, DIR_BG, DIR_FG, DIR_SYS):
+		img_resize(DIR_BG, DIR_FG, DIR_SYS, DIR_IMG)
+		text_dec(DIR_SCR_DEC, DIR_SCR, EXE_IGS)
+		text_cnv(DEFAULT_TXT, DIR_SCR_DEC, ZERO_TXT)
+		junk_del(DIR_BG, DIR_FG, DIR_SYS, DIR_SCR, DIR_SCR_DEC, DEFAULT_TXT, debug)
+
+
+main()
